재난실행 절차 수정  건물 선택 모달 구현

+++ src/views/subs/SopList.vue
@@ -11,12 +11,12 @@
           template(slot="condition1", slot-scope="props")
             select.ui.dropdown(v-model="req.msfrtnKndCd")
               option(value="") 재난종류
-              option(v-for="code in typeCode", value="code.cmmnCd") {{code.cmmnCdNm}}
+              option(v-for="code in typeCode", :value="code.cmmnCd") {{code.cmmnCdNm}}
 
           template(slot="condition2", slot-scope="props")
             select.ui.dropdown(v-model="req.crisisGnfdStepCd")
               option(value="") 위기발령단계
-              option(v-for="code in stepCode", value="code.cmmnCd") {{code.cmmnCdNm}}
+              option(v-for="code in stepCode", :value="code.cmmnCd") {{code.cmmnCdNm}}
       div.sub-content
         div.content
           DataTable(
@@ -48,16 +48,17 @@
               template(v-if="getUser.userCode == 'S0400100'")
                 button.ui.button.large() 편집
                 button.ui.button.large(@click="deleteSop") 삭제
-                button.ui.button.large.green SOP실행
+              button.ui.button.large.green(@click="selectMode") SOP실행
             div.btn-wrap.right
               template(v-if="getUser.userCode == 'S0400100'")
-              button.ui.button.large.blue 재난절차생성
+              button.ui.button.large.blue(@click="$router.push({name: 'sop-edit'})") 재난절차생성
 
 </template>
 
 <script>
 import DataTable from '@/components/DataTable'
 import SearchComp from '@/components/SearchComp'
+import SelectSopMode from '@/components/SelectSopMode'
 import { sopListHeader } from '@/setting'
 import SopManageApi from '@/api/SopManage'
 import { mapGetters, mapActions } from 'vuex'
@@ -86,12 +87,14 @@ export default {
         sopListData: []
       },
       typeCode: [],
-      stepCode: []
+      stepCode: [],
+      isSelected: false
     }
   },
   components: {
     DataTable,
-    SearchComp
+    SearchComp,
+    SelectSopMode
   },
   created() {
     this.getCodeList('S090')
@@ -127,17 +130,20 @@ export default {
       })
     },
     deleteSop () {
-      const requestData = JSON.stringify({
-        sopId: this.sopList.selected[0].sopId
-      })
-      SopManageApi.deleteItem(requestData).then(result => {
-        console.log(result.data)
-        this.getSopList()
-      }).catch(error => {
-        const err = error.response
-        console.log(err)
-        this.$modal.show('dialog', codeGenerator(err.data.msgCode, err.data.msgValue))
-      })
+      this.selectCheck()
+      if(this.isSelected) {
+        const requestData = JSON.stringify({
+          sopId: this.sopList.selected[0].sopId
+        })
+        SopManageApi.deleteItem(requestData).then(result => {
+          console.log(result.data)
+          this.getSopList()
+        }).catch(error => {
+          const err = error.response
+          console.log(err)
+          this.$modal.show('dialog', codeGenerator(err.data.msgCode, err.data.msgValue))
+        })
+      }
     },
     getCodeList (code) {
       const requestData = JSON.stringify({
@@ -161,6 +167,29 @@ export default {
       this.sopList.selected = []
       if(!itemInfo.selected) {
         this.sopList.selected.push(this.sopList.sopListData[itemInfo.idx])
+        this.isSelected = true
+      }
+    },
+    selectMode () {
+      this.selectCheck()
+      if(this.isSelected) {
+        this.$modal.show(SelectSopMode,{
+          title: '실행모드',
+          data: this.sopList.selected[0]
+        },{
+          height: 'auto'
+        })
+      }
+    },
+    selectCheck () {
+      if(this.sopList.selected.length == 0) {
+        this.isSeleced = false
+        this.$modal.show('dialog', {
+          title: '선택오류',
+          text: '선택 된 SOP가 없습니다.'
+        })
+      } else {
+        this.isSelected = true
       }
     }
   }
